requirements.txt
google-generativeai
PyYAML
python-dotenv
Flask
cryptography
